# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Totem Bridge Off Chain Service : This is a Bridge server.",
        "title": "Totem Bridge Off Chain API",
        "contact": {},
        "version": "1.0"
    },
    "host": "dev-totem-bridge-offchain-lidbscqqia-uc.a.run.app",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/withdraw_token": {
            "post": {
                "description": "The With Draw Token API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "WithdrawToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain",
                        "name": "chain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "TokenAddr",
                        "name": "token_addr",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Recipient",
                        "name": "recipient",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Amount",
                        "name": "amount",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawTokenResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponseDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrorResponseDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.WithdrawTokenResponseDto": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}